backend/requirements.txt
flask~=3.1.1
werkzeug~=3.1.3
opencv-python~=4.11.0.86
numpy~=1.24.4
python-dotenv~=1.1.0
huggingface_hub
flask_cors
flask-bcrypt
bcrypt
flask-jwt-extended
flask_sqlalchemy
SQLAlchemy~=2.0.41
psycopg2-binary
flask-socketio
eventlet
boto3~=1.38.31
botocore~=1.38.31
websocket~=0.2.1
imageio[pyav]~=2.37.0
imageio[ffmpeg]
pillow


db~=0.1.1
requests~=2.32.3
python-dateutil~=2.9.0.post0
torch~=2.7.0
einops~=0.8.1